requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
pydantic~=1.8.2
requests==2.26.0
sniffio==1.2.0
SQLAlchemy~=1.4.28
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
bcrypt==3.2.0
cryptography~=36.0.1
passlib~=1.7.4
python-jose==3.3.0
python-multipart==0.0.5
sphinx==4.4.0
redis==4.1.4
rejson==0.5.6
pylint==2.12.2
flower==1.0.0
celery==5.2.3